+++ fastapi_restful/__init__.py
@@ -1,5 +1,5 @@
-from .api import APIVersion, RestAPI
+from .api import RestAPIRouter, RestAPI
 from .resource import Resource
 from .utils import route_settings
 
-__all__ = ["RestAPI", "Resource", "route_settings", "APIVersion"]
+__all__ = ["RestAPI", "Resource", "route_settings", "RestAPIRouter"]