+++ tests/test_api.py
@@ -8,14 +8,14 @@ from fastapi_restful import RestAPI
 
 def test__init_rest_api__success():
     app = FastAPI()
-    api = RestAPI(fastapi_app=app)
+    api = RestAPI()
     assert api._fastapi is app
     assert api.router.prefix == "/api"
 
 
 def test__init_rest_api__empty_default_prefix__success():
     app = FastAPI()
-    api = RestAPI(fastapi_app=app, prefix=None)
+    api = RestAPI(path=None)
     assert api._fastapi is app
     assert not hasattr(api.router, "prefix")
 
@@ -28,7 +28,7 @@ def test__add_resource__success(resource_mock, rest_api_instance):
 def test__get_item__success(rest_api_instance, api_version_instance):
     assert rest_api_instance.versions == {}
     prefix = api_version_instance.router.prefix
-    rest_api_instance._versions[prefix] = api_version_instance
+    rest_api_instance.rest_api_routers[prefix] = api_version_instance
     assert rest_api_instance[prefix] is api_version_instance
 
 
@@ -54,7 +54,7 @@ def test__apply__success(rest_api_instance, resource_type, api_version_instance)
 
     resource_type.get = get
     api_version_instance.add_resource(resource_type)
-    rest_api_instance.include_api_version(api_version_instance)
+    rest_api_instance.include_rest_api_router(api_version_instance)
     assert rest_api_instance.apply() is None
 
 
@@ -67,6 +67,6 @@ def test__call_http_methods__success(resource_type):
 
 def test__include_api_version__failed__prefix_is_exist(rest_api_instance, api_version_instance):
     expected_error_msg = "This version is exist: /v1"
-    rest_api_instance._versions[api_version_instance.router.prefix] = api_version_instance
-    exc = pytest.raises(AssertionError, rest_api_instance.include_api_version, api_version_instance)
+    rest_api_instance.rest_api_routers[api_version_instance.router.prefix] = api_version_instance
+    exc = pytest.raises(AssertionError, rest_api_instance.include_rest_api_router, api_version_instance)
     assert str(exc.value) == expected_error_msg